風発生の時刻指定可能
