+++ Assets/Scenes 1/AddWind.unity
@@ -157,9 +157,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   coefficient: 3
-  velocity: {x: 0, y: 5, z: 5}
+  velocity: {x: 0, y: 2, z: 5}
   rb: {fileID: 0}
-  isTap: 0
+  start: 0
+  end: 0
 --- !u!54 &89143553
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -379,9 +380,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   coefficient: 3
-  velocity: {x: 0, y: 5, z: 5}
+  velocity: {x: 0, y: 2, z: 5}
   rb: {fileID: 0}
-  isTap: 0
+  start: 5
+  end: 8
 --- !u!1 &588232391
 GameObject:
   m_ObjectHideFlags: 0
@@ -458,13 +460,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 588232391}
-  m_LocalRotation: {x: -0.09067106, y: 0.7131859, z: -0.09389268, w: -0.6887153}
-  m_LocalPosition: {x: 15.62, y: 5.51, z: 0.81}
+  m_LocalRotation: {x: 0, y: -0.7071068, z: 0, w: 0.7071068}
+  m_LocalPosition: {x: 16.58, y: 3.6, z: -0.25}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 15, y: 268, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: -90, z: 0}
 --- !u!1 &807581871
 GameObject:
   m_ObjectHideFlags: 0
@@ -602,7 +604,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
       propertyPath: m_LocalRotation.w
-      value: 1
+      value: 0.9990483
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
       propertyPath: m_LocalRotation.x
@@ -610,7 +612,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
       propertyPath: m_LocalRotation.y
-      value: 0
+      value: 0.043619405
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
       propertyPath: m_LocalRotation.z
@@ -622,7 +624,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
@@ -634,3 +636,32 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
+--- !u!1 &2085718114 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 21594614f6ff142489b6dd47705272d2, type: 3}
+  m_PrefabInstance: {fileID: 1976285421}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &2085718115
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2085718114}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 94c1b0262dbc4614498474d4c13fccd9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &2085718116
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2085718114}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3f689c70ce8330d428848321ac552ecd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
